+++ Acornote_Kotlin/app/src/main/java/tonnysunm/com/acornote/model/LabelDao.kt
@@ -15,7 +15,7 @@ interface LabelDao : BaseDao<Label> {
     fun getLabelsWithNoteCount(): LiveData<List<LabelWithNoteCount>>
 
     @Query("SELECT a.id, a.title, count(b.id) > 0 AS checked, :noteId AS noteId FROM label_table a LEFT JOIN note_label_table b ON a.id = b.label_id AND b.note_id = :noteId GROUP BY a.id ORDER BY a.created_at DESC")
-    fun getLabelsWithNoteId(noteId: Long): DataSource.Factory<Int, LabelWithCheckStatus>
+    fun getLabelsWithNoteId(noteId: Int): DataSource.Factory<Int, LabelWithCheckStatus>
 
 //    @Query("SELECT id, title, 0 as checked, 0 as noteId FROM label_table ORDER BY created_at DESC")
 //    fun getPagingAll(): DataSource.Factory<Int, LabelWithCheckStatus>
@@ -24,7 +24,7 @@ interface LabelDao : BaseDao<Label> {
     suspend fun getAll(): List<Label>
 
     @Query("SELECT * from label_table WHERE id = :id LIMIT 1")
-    fun getLabel(id: Long): LiveData<Label>
+    fun getLabel(id: Int): LiveData<Label>
 
     @Insert
     suspend fun insert(entities: List<Label>)
